enhancement: move disco to runtime package
