+++ cmd/disco/runtime/flags.go
@@ -3,7 +3,7 @@
 // Distributed under the GNU Lesser General Public License v3.0 software license, see the accompanying
 // file LICENSE or <https://www.gnu.org/licenses/lgpl-3.0.html>
 
-package main
+package runtime
 
 import (
 	"github.com/vechain/thor/v2/log"
